Move blog index template to src/templates.
Having it in src/pages was causing issues because Gatsby was attempting
to treat it as a static page rather than a dynamic one, but the GraphQL
query for the page only works when extra page context variables are
added.
